pyproject.toml
[tool.poetry]
name = "mattes-allegro-pl"
version = "0.7.1"
description = "Python client for Allegro.pl API"
authors = ["Raphael Krupinski"]
license = 'MIT'
packages=[{include='allegro_pl', from='src'}]
readme = 'README.rst'
repository = 'https://github.com/mattesilver/allegro-pl/'

[tool.poetry.dependencies]
python = "^3.7"
allegro-pl-rest-api = "^2019.3"
oauthlib = "^3.0"
requests-oauthlib = "^1.2"
tenacity = "^5.0.4"
zeep = "^3.3"

[tool.poetry.dev-dependencies]
codecov = "^2.0"
pytest = "^5.3"
pytest-cov = "^2.6"
toml = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
